moved matter-js stuff to service file
