+++ lib/blastronauts/src/app/black-hole/black-hole.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import * as Matter from 'matter-js';
+import { MatterService } from '../services/matter.service';
 
 @Component({
   selector: 'app-black-hole',
@@ -8,38 +8,10 @@ import * as Matter from 'matter-js';
 })
 export class BlackHoleComponent implements OnInit {
 
-  // matter-js aliases
-  private Engine = Matter.Engine;
-  private Bodies = Matter.Bodies;
-  private Render = Matter.Render;
-  private World = Matter.World;
-
-  private engine;
-  private renderer;
-
-  constructor() {
-    this.engine = this.Engine.create();
-    this.renderer = this.Render.create({
-      element: document.body,
-      engine: this.engine
-    });
-
-    // create two boxes and a ground
-    var boxA = this.Bodies.rectangle(400, 200, 80, 80);
-    var boxB = this.Bodies.rectangle(450, 50, 80, 80);
-    var ground = this.Bodies.rectangle(400, 610, 810, 60, { isStatic: true });
-
-    // add all of the bodies to the world
-    this.World.add(this.engine.world, [boxA, boxB, ground]);
-
-    // run the engine
-    this.Engine.run(this.engine);
-
-    // run the renderer
-    this.Render.run(this.renderer);
+  constructor(private matterService: MatterService) {
+    this.matterService.run();
   }
 
-  ngOnInit(): void {
-  }
+  ngOnInit(): void {}
 
 }
